+++ src/client/app/actions/index.js
@@ -17,3 +17,29 @@ export const toggleTodo = (id) => {
     id
   };
 };
+
+export const fetchingTodos = () => {
+  return {
+    type: t.FETCHING_TODOS
+  };
+};
+
+export const fetchedTodos = (todos) => {
+  return {
+    type: t.FETCHED_TODOS,
+    todos
+  };
+};
+
+export const fetchTodos = () => {
+  return (dispatch) => {
+    dispatch(fetchingTodos());
+
+    window.fetch('/api/todos')
+      .then((response) => response.json())
+      .then((todos) => {
+        dispatch(fetchedTodos(todos));
+        todos.forEach(todo => dispatch(addTodo(todo)));
+      });
+  };
+};